requirements.txt
# AI Quant Trading System - Requirements
# Python 3.11+

# Core
asyncio-throttle>=1.0.2
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Data Processing
pandas>=2.1.0
numpy>=1.26.0
scipy>=1.11.0

# Database
asyncpg>=0.29.0
sqlalchemy>=2.0.0
redis>=5.0.0

# Exchange & Trading
ccxt>=4.1.0

# Web & API
fastapi>=0.104.0
uvicorn>=0.24.0
aiohttp>=3.9.0

# Task Scheduling
croniter>=2.0.0

# Machine Learning
scikit-learn>=1.3.0

# Statistics (for cointegration tests)
statsmodels>=0.14.0

# System Monitoring
psutil>=5.9.0

# Data Formats
pyyaml>=6.0.0

# Logging
structlog>=23.2.0

# HTTP Client
httpx>=0.25.0

# Testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Type Checking (optional)
mypy>=1.7.0

# Streamlit Dashboard
streamlit>=1.28.0
plotly>=5.18.0
httpx>=0.25.0
